+++ pyawsopstoolkit/insights/iam/__main__.py
@@ -2,22 +2,18 @@ import re
 from datetime import datetime
 from typing import Optional
 
-import pyawsopstoolkit.models
 from pyawsopstoolkit.__interfaces__ import ISession
 from pyawsopstoolkit.__validations__ import Validation
 
 
-class IAM:
+class Role:
     """
-    A class that encapsulates insights related to the IAM service, including roles, users, and other entities.
+    A class representing insights related with IAM roles.
     """
 
-    def __init__(
-            self,
-            session: ISession
-    ) -> None:
+    def __init__(self, session: ISession) -> None:
         """
-        Initializes the constructor of the IAM class.
+        Initializes the constructor of the Role class.
         :param session: An ISession object providing access to AWS services.
         :type session: ISession
         """
@@ -49,7 +45,7 @@ class IAM:
             self,
             no_of_days: Optional[int] = 90,
             include_newly_created: Optional[bool] = False
-    ) -> list[pyawsopstoolkit.models.iam.role.Role]:
+    ) -> list:
         """
         Returns a list of unused IAM roles based on the specified parameters.
         :param no_of_days: The number of days (integer) to check if the IAM role has been used within the
@@ -61,14 +57,14 @@ class IAM:
         :return: A list of unused IAM roles.
         :rtype: list
         """
+        from pyawsopstoolkit.advsearch.iam import Role
+
         Validation.validate_type(no_of_days, int, 'no_of_days should be an integer.')
         Validation.validate_type(include_newly_created, bool, 'include_newly_created should be a boolean.')
 
-        from pyawsopstoolkit import advsearch
-
         current_date = datetime.today().replace(tzinfo=None)
-        iam_object = advsearch.IAM(self.session)
-        iam_roles = iam_object.search_roles(include_details=True)
+        role_object = Role(self.session)
+        iam_roles = role_object.search_roles(include_details=True)
 
         if iam_roles is None:
             return []
@@ -98,11 +94,47 @@ class IAM:
 
         return unused_roles_list
 
+
+class User:
+    """
+    A class representing insights related with IAM users.
+    """
+
+    def __init__(self, session: ISession) -> None:
+        """
+        Initializes the constructor of the User class.
+        :param session: An ISession object providing access to AWS services.
+        :type session: ISession
+        """
+        Validation.validate_type(session, ISession, 'session should be of Session type.')
+
+        self._session = session
+
+    @property
+    def session(self) -> ISession:
+        """
+        Gets the ISession object which provides access to AWS services.
+        :return: The ISession object which provide access to AWS services.
+        :rtype: ISession
+        """
+        return self._session
+
+    @session.setter
+    def session(self, value: ISession) -> None:
+        """
+        Sets the ISession object which provides access to AWS services.
+        :param value: The ISession object which provides access to AWS services.
+        :type value: ISession
+        """
+        Validation.validate_type(value, ISession, 'session should be of Session type.')
+
+        self._session = value
+
     def unused_users(
             self,
             no_of_days: Optional[int] = 90,
             include_newly_created: Optional[bool] = False
-    ) -> list[pyawsopstoolkit.models.iam.user.User]:
+    ) -> list:
         """
         Returns a list of unused IAM users based on the specified parameters.
         :param no_of_days: The number of days (integer) to check if the IAM user has been used within the
@@ -114,14 +146,14 @@ class IAM:
         :return: A list of unused IAM users.
         :rtype: list
         """
+        from pyawsopstoolkit.advsearch.iam import User
+
         Validation.validate_type(no_of_days, int, 'no_of_days should be an integer.')
         Validation.validate_type(include_newly_created, bool, 'include_newly_created should be a boolean.')
 
-        from pyawsopstoolkit import advsearch
-
         current_date = datetime.today().replace(tzinfo=None)
-        iam_object = advsearch.IAM(self.session)
-        iam_users = iam_object.search_users(include_details=True)
+        user_object = User(self.session)
+        iam_users = user_object.search_users(include_details=True)
 
         if iam_users is None:
             return []