+++ pyawsopstoolkit/security/__init__.py
@@ -1,5 +1,5 @@
 __all__ = [
-    "IAM"
+    "iam"
 ]
 __name__ = "pyawsopstoolkit.security"
 __description__ = """
@@ -8,5 +8,3 @@ associated with AWS (Amazon Web Services), such as IAM roles without permissions
 these features are finely tuned to meet the unique demands of the expansive AWS ecosystem, encompassing a diverse
 spectrum of aspects.
 """
-
-from pyawsopstoolkit.security.__main__ import IAM